Move all db interaction to domain to simplify development
