+++ src/Common/Domain/Context/ISeeder.cs
@@ -1,6 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 
-namespace Trace.Common.Domain;
+namespace Trace.Common.Domain.Context;
 
 public interface ISeeder {
     DateTimeOffset Time => DateTimeOffset.Now;
