+++ src/Common/Infrastructure/Trace.Common.Infrastructure.csproj
@@ -9,12 +9,10 @@
   <ItemGroup>
     <Folder Include="FIles\" />
     <Folder Include="Identity\" />
-    <Folder Include="Persistence\Migrations\" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="CassandraCSharpDriver" Version="3.19.1" />
-    <PackageReference Include="EFCore.NamingConventions" Version="7.0.2" />
     <PackageReference Include="Hangfire.AspNetCore" Version="1.8.0-rc2" />
     <PackageReference Include="Hangfire.Dashboard.Basic.Authentication" Version="5.0.0" />
     <PackageReference Include="Hangfire.Redis.StackExchange" Version="1.9.1-beta" />
@@ -28,21 +26,10 @@
     <PackageReference Include="HotChocolate.Stitching.Redis" Version="13.0.5" />
     <PackageReference Include="HotChocolate.Subscriptions.Redis" Version="13.0.5" />
     <PackageReference Include="HotChocolate.Types" Version="13.0.5" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.4" />
     <PackageReference Include="Microsoft.AspNetCore.DataProtection.StackExchangeRedis" Version="7.0.3" />
     <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="7.0.3" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.4">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.4">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
     <PackageReference Include="Minio" Version="4.0.7" />
     <PackageReference Include="MQTTnet.AspNetCore" Version="4.1.4.563" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="7.0.3" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL.NetTopologySuite" Version="7.0.3" />
     <PackageReference Include="Steeltoe.Bootstrap.Autoconfig" Version="3.2.3" />
     <PackageReference Include="Steeltoe.Connector.ConnectorCore" Version="3.2.3" />
     <PackageReference Include="Steeltoe.Discovery.ClientCore" Version="3.2.3" />
